+++ src/utils/FormJson.tsx
@@ -1,0 +1,9 @@
+export declare module FormJson {
+  export interface Element {
+    name: string;
+    type: string;
+  }
+  export interface Root {
+    elements: Element[];
+  }
+}
